# small idct code improvement